Add working version of text file layer. (For those of you playing along at home, this was about 45 minutes of work -- part of which was trying to figure out why the data file wouldn't work, only to find out I had expandtab on in vim -- so I wasn't typing tabs, i was typing spaces!)
This has uncovered the fact that there is some logical error in our code somewhere: Visit markers.html then click the purple marker repeatedly.  Despite the fact that you start clicking in the sea, the marker moves north, as does your center point.

This text file layer does not yet support popups or anything else, just drawing the images at a lat/lon with an 'image' given in the textfile.


git-svn-id: http://svn.openlayers.org/trunk/openlayers@172 dc9f47b5-9b13-0410-9fdd-eb0c1a62fdaf