No animation frames on hidden windows.
As an optimization, Firefox does not execute functions passed to requestAnimationFrame when the window is hidden.  This is the case for the panTo tests here that use a map in a hidden iframe.  Since other browsers may follow suit, we just test the fallback setTimeout method of animating on these tests.