Fix OpenLayers.Layer and OpenLayers.Map apidoc
'viewPortPx' parameter for OpenLayers.Layer.getLonLatFromViewPortPx
can be either a Pixel or a simple javascript object. 'lonlat'
parameter for OpenLayers.Layer.getViewPortPxFromLonLat can be either a
LonLat or a simple javascript object.

'px' parameter for OpenLayers.Map.getLonLatFromViewPortPx and
OpenLayers.Map.getLonLatFromPixel can be either a Pixel or a simple
javascript object.