Making map construction work with permalink.
If a map is configured with a permalink control and a center, the permalink control tries to work with the map center before setCenter is called.  This sequence assumes that the map center will already be a LonLat.  This fails if the map is constructed with a coordinate array.  Safer move is to convert center to LonLat before controls are added.