Fixing the extend method to work on IE.
In IE, events don't have a `hasOwnProperty` method.  Instead of trusting that the source objects have this method, we call the same on the Object prototype.  If we want to be safer about the `hasOwnProperty` calling done in e3cc96dbfb58ee873db457e74de7c7774a25bf45, we could create an `OpenLayers.Object.has` method that also used the Object prototype.