Making testing in IE a tiny bit less painful.
The super handy quick filter that Marc added in 57c16b73458b3c16d28695b74f99bdda38236da5 was causing a bit of trouble in IE.  First, `input` was undefined where it was used as a misspelling for `window`.  After that, IE choked on the nice CSS3 selector when filtering the test list.  It turns out we don't need to be as selective and can treat all anchor elements within the table rows as candidates.