A simple zoom control with zoom in/out links.
The zoom control allows for zoom in/out links that can be styled with CSS.

Note: This change was originally captured in #269.  It involves nice additions thanks @ahocevar.  The changes were unintentionally merged and then reverted with fb3caf1561f9db9c4516b12cd712217a04910cf8, so the history of commits is not immediately apparent (though still likely there somewhere due to the magic of git).