Allow for circular dependency in XMLHttpRequest.js
To make custom builds easier, it's convenient to have a OpenLayers.Request depend on OpenLayers.Request.XMLHttpRequest (see 18b4d6808da1278e9ec43caba56281804f82e63e).  This makes it so we don't know which is going to be included first.  To get around this, we can be careful about when we define OpenLayers.Request.