Leaving animation entirely up to the renderers
With this change, the map knows nothing of animation.  Different renderers
may behave differently when their render method is called.  A PDF renderer
(thought experiment at this point) might immediately render to a file.  The
dom based renderers schedule a repaint for the next animation frame or in a
timeout depending on capabilities - in either case, setting multiple map
properties will *not* result in multiple reflow/repaints.