Accept a render callback for state generation
The map's render method can accept a callback that serves as a state generator.  This callback will be called before rendering.  Any changes to the map state can be made before rendering takes place (multiple changes will not result in multiple renders).  The return from this callback indicates whether the generator is exhausted.