(Re)add a reallylean target to the build.
Previously we had both a `clean` and `reallyclean` in our Makefile; both
got lost when we switched to a python build.

A clean 'target' can be invoked by calling the build script with the '-c' flag:
  ./build.py -c

This commit readds the `reallyclean` target and lets it clean up a little more
than the Makefiles `clean` or `reallyclean` did. More files are deleted,
but only if these are being ignored inside of the repository (e.g. through
.gitignore files).